Fix negation of float literals

+++ src/test/run-pass/issue-7222.rs
@@ -11,10 +11,10 @@
 // pretty-expanded FIXME #23616
 
 pub fn main() {
-    const FOO: f64 = 10.0;
+    const FOO: u64 = 10;
 
-    match 0.0 {
-        0.0 ... FOO => (),
+    match 0 {
+        0 ... FOO => (),
         _ => ()
     }
 }
